Add unit tests for raw_usb backend
