+++ tests/backends/hardware/j5/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for j5 hardware implementations."""
